Fix coin score value

+++ Penguin/Config.swift
@@ -10,7 +10,7 @@ enum Config {
     /// Base score added based on the distance
     static let baseScore: Double = 1
     /// Base coin score (increased according to the speed)
-    static let baseCoinValue: Int = 20
+    static let baseCoinValue: Double = 20
 
     /// Maximum rotation angle of the player (in the edges), in degrees
     static let maxRotationAngle: Double = 15
